Convert SaveFile function private to public
