+++ eCommerce.BackendApi/Interfaces/IFileStorageService.cs
@@ -3,9 +3,9 @@ namespace eCommerce.BackendApi.Interfaces
 {
 	public interface IFileStorageService
 	{
-		string GetFileUrl(string fileName);
 		Task SaveFileAsync(Stream mediaBinaryStream, string fileName);
 		Task DeleteFileAsync(string fileName);
+		Task<string> SaveFile(IFormFile file);
 	}
 }
